refactoring code for duplication
